:recycle: [refactory]: Divide Component
- Divide CollectionBody (parent)
- Add CollectionList (child)
- Add CollectionControl (child)
