+++ src/pages/collections/application/CollectionList.tsx
@@ -1,196 +1,42 @@
-import { FormProvider, useForm } from 'react-hook-form';
-import 'styled-components';
 import styled from 'styled-components';
 
-import Loadings from '../../../components/Loadings';
-import { thisIdExcept } from '../../../utils/config';
-import { colors } from '../../../utils/config';
+import { colors, thisIdExcept } from '../../../utils/config';
 import { media, noDrag, scroll } from '../../../utils/mixins';
-import { useCollectionDel } from '../custom_hooks/useCollectionDel';
-import { useCollectionFetch } from '../custom_hooks/useCollectionFetch';
-import { useCollectionsDel } from '../custom_hooks/useCollectionsDel';
-import { CollectionNamesDTO } from '../dto/dto';
+import { CollectionDTO } from '../dto/dto';
 import InputCheck from './InputCheck';
 
-function CollectionList() {
-  const { isCollectLoading, isCollectFetching, collectData, collectLength } = useCollectionFetch();
-  const { delLoading, delData } = useCollectionDel();
-  const { delsLoading, delDatas } = useCollectionsDel();
-  const isLoading = isCollectLoading || isCollectFetching;
-  const isFetcing = delLoading || delsLoading;
-  const method = useForm<CollectionNamesDTO>();
-  const checkCount = method.watch().names ? method.watch().names.length : 0;
+interface ICollectionList {
+  isLoading: boolean;
+  collections: CollectionDTO[] | [] | undefined;
+  onDelete: (name: string) => void;
+}
 
-  const onClick = (name: string) => {
-    const data = name;
-    if (thisIdExcept(data)) {
-      alert('It cannot be done.');
-      return;
-    }
-    const check = window.confirm(`Delete the [ ${data} ] collection?`);
-    if (check) {
-      delLoading || delData(data, {
-        onSuccess: () => method.reset()
-      });
-    }
-  }
-
-  const onSubmit = ({ names }: CollectionNamesDTO) => {
-    const datas = names;
-    if (!datas || !datas.length) {
-      return;
-    }
-    const check = window.confirm(`Delete the [ ${datas.length} ] collections?`);
-    if (check) {
-      delsLoading || delDatas(datas, {
-        onSuccess: () => method.reset()
-      });
-    }
-  }
-
-  const onReset = () => {
-    method.reset();
-  }
-
+function CollectionList({ isLoading, collections, onDelete }: ICollectionList) {
+  
   return (
-    <>
-      <Section>
-        <FormProvider {...method}>
-          <form onSubmit={method.handleSubmit(onSubmit)}>
-            <Head>
-              <Title>
-                Collections
-                <Count>{ `Total: ${collectLength}` }</Count>
-              </Title>
-              <Control>
-                <Reset type="button" onClick={onReset}>Reset</Reset>
-                <Deletes type="submit" chk={ checkCount }>
-                  <img src="./assets/svg/trash-can-solid.svg" alt="delete all" />
-                  <CheckNum>{ checkCount }</CheckNum>
-                </Deletes>
-              </Control>
-            </Head>
-            <Contain>
-              {
-                isLoading ?
-                  <Loading>Loading...</Loading> :
-                  collectData?.map((item) => (
-                    <li key={item.name}>
-                      {thisIdExcept(item.name) ?
-                        <NotCheck /> :
-                        <InputCheck thisName={item.name} />}
-                      <Wrapper>
-                        <Name>{item.name}</Name>
-                        <Type>{item.type}</Type>
-                      </Wrapper>
-                      {thisIdExcept(item.name) ||
-                        <Delete type="button" onClick={() => onClick(item.name)}>Delete</Delete>}
-                    </li>  
-                ))
-              }
-            </Contain>
-          </form>
-        </FormProvider>
-      </Section>
-
-      {isFetcing && <Loadings />}
-    </>
+    <Contain>
+      {
+        isLoading ?
+          <Loading>Loading...</Loading> :
+          collections?.map((item) => (
+            <li key={item.name}>
+              {thisIdExcept(item.name) ?
+                <NotCheck /> :
+                <InputCheck thisName={item.name} />}
+              <Wrapper>
+                <Name>{item.name}</Name>
+                <Type>{item.type}</Type>
+              </Wrapper>
+              {thisIdExcept(item.name) ||
+                <Delete type="button" onClick={() => onDelete(item.name)}>Delete</Delete>}
+            </li>  
+        ))
+      }
+    </Contain>
   )
 }
 
 export default CollectionList;
-
-const Section = styled.section`
-  padding: 20px 0;
-`;
-
-const Head = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const Title = styled.h2`
-  font-size: 30px;
-  font-weight: 600;
-  margin-bottom: 20px;
-`;
-
-const Count = styled.strong`
-  font-size: 18px;
-  font-weight: 600;
-  color: ${colors.secondary};
-  padding: 0 20px;
-
-  @media ${media.tablet_s} {
-    font-size: 25px;
-  }
-`;
-
-const Control = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: stretch;
-
-  @media ${media.tablet_s} {
-    flex-direction: row;
-  }
-`;
-
-const Reset = styled.button`
-  font-size: 16px;
-  color: ${colors.light};
-  background-color: ${colors.green};
-  border-radius: 5px;
-  padding: 5px 10px;
-  margin-bottom: 10px;
-  cursor: pointer;
-  transition: opacity 200ms ease-in-out;
-
-  &:active {
-    opacity: .4;
-  }
-
-  @media ${media.tablet_s} {
-    margin: 0 20px;
-    &:hover {
-      opacity: .4;
-    }
-  }
-`;
-
-const Deletes = styled.button<{chk: number}>`
-  position: relative;
-  width: 30px;
-  padding: 5px;
-  cursor: pointer;
-  transition: opacity 200ms ease-in-out;
-  opacity: ${({ chk }) => chk === 0 ? .4 : 1};
-
-  &:active {
-    opacity: .4;
-  }
-
-  @media ${media.tablet_s} {
-    margin-right: 30px;
-    &:hover {
-      opacity: .4;
-    }
-  }
-`;
-
-const CheckNum = styled.sub`
-  position: absolute;
-  right: -22px;
-  top: 2px;
-  width: 25px;
-  height: 30px;
-  border-radius: 5px;
-  color: ${colors.light};
-  font-size: 20px;
-  font-weight: 600;
-`;
 
 const Contain = styled.ul`
   display: flex;
@@ -288,5 +134,3 @@ const NotCheck = styled.div`
   margin: 10px;
   align-self: flex-start;
 `;
-
-
