+++ src/pages/collections/application/InputCheck.tsx
@@ -1,5 +1,6 @@
-import { useFormContext } from 'react-hook-form';
 import styled from 'styled-components';
+import { useFormContext } from 'react-hook-form';
+
 import { colors } from '../../../utils/config';
 import { media } from '../../../utils/mixins';
