+++ src/pages/login/application/LoginId.tsx
@@ -1,9 +1,9 @@
-import { useFormContext } from 'react-hook-form';
 import styled from 'styled-components';
+import { useFormContext } from 'react-hook-form';
+
 import { colors } from '../../../utils/config';
 
 function LoginId() {
-
   const { register, formState: { errors } } = useFormContext<{admin_id: string}>();
   
   return (
